refactor: improved the search students by identifier modal
